+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/account/AccountTestSetup.kt
@@ -4,6 +4,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import co.electriccoin.zcash.ui.common.model.WalletSnapshot
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
+import co.electriccoin.zcash.ui.screen.account.history.fixture.TransactionHistorySyncStateFixture
+import co.electriccoin.zcash.ui.screen.account.state.TransactionHistorySyncState
 import co.electriccoin.zcash.ui.screen.account.view.Account
 import java.util.concurrent.atomic.AtomicInteger
 
@@ -11,10 +13,26 @@ class AccountTestSetup(
     private val composeTestRule: ComposeContentTestRule,
     private val walletSnapshot: WalletSnapshot,
 ) {
+    // TODO [#1282]: Update AccountView Tests #1282
+    // TODO [#1282]: https://github.com/Electric-Coin-Company/zashi-android/issues/1282
+
+    val initialHistorySyncState: TransactionHistorySyncState = TransactionHistorySyncStateFixture.new()
+
     private val onSettingsCount = AtomicInteger(0)
     private val onReceiveCount = AtomicInteger(0)
     private val onSendCount = AtomicInteger(0)
-    private val onHistoryCount = AtomicInteger(0)
+    private val onItemClickCount = AtomicInteger(0)
+    private val onItemIdClickCount = AtomicInteger(0)
+
+    fun getOnItemClickCount(): Int {
+        composeTestRule.waitForIdle()
+        return onItemClickCount.get()
+    }
+
+    fun getOnItemIdClickCount(): Int {
+        composeTestRule.waitForIdle()
+        return onItemIdClickCount.get()
+    }
 
     fun getOnSettingsCount(): Int {
         composeTestRule.waitForIdle()
@@ -29,11 +47,6 @@ class AccountTestSetup(
     fun getOnSendCount(): Int {
         composeTestRule.waitForIdle()
         return onSendCount.get()
-    }
-
-    fun getOnHistoryCount(): Int {
-        composeTestRule.waitForIdle()
-        return onHistoryCount.get()
     }
 
     fun getWalletSnapshot(): WalletSnapshot {
@@ -51,9 +64,13 @@ class AccountTestSetup(
                 onSettingsCount.incrementAndGet()
             },
             goBalances = {},
-            goHistory = {
-                onHistoryCount.incrementAndGet()
+            transactionState = initialHistorySyncState,
+            onItemClick = {
+                onItemClickCount.incrementAndGet()
             },
+            onTransactionIdClick = {
+                onItemIdClickCount.incrementAndGet()
+            }
         )
     }