+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/account/history/HistoryTestSetup.kt
@@ -1,23 +1,17 @@
-package co.electriccoin.zcash.ui.screen.history
+package co.electriccoin.zcash.ui.screen.account.history
 
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
-import co.electriccoin.zcash.ui.screen.history.state.TransactionHistorySyncState
-import co.electriccoin.zcash.ui.screen.history.view.History
+import co.electriccoin.zcash.ui.screen.account.state.TransactionHistorySyncState
+import co.electriccoin.zcash.ui.screen.account.view.HistoryContainer
 import java.util.concurrent.atomic.AtomicInteger
 
 class HistoryTestSetup(
     private val composeTestRule: ComposeContentTestRule,
     initialHistorySyncState: TransactionHistorySyncState
 ) {
-    private val onBackClickCount = AtomicInteger(0)
     private val onItemClickCount = AtomicInteger(0)
     private val onItemIdClickCount = AtomicInteger(0)
-
-    fun getOnBackClickCount(): Int {
-        composeTestRule.waitForIdle()
-        return onBackClickCount.get()
-    }
 
     fun getOnItemClickCount(): Int {
         composeTestRule.waitForIdle()
@@ -32,11 +26,8 @@ class HistoryTestSetup(
     init {
         composeTestRule.setContent {
             ZcashTheme {
-                History(
+                HistoryContainer(
                     transactionState = initialHistorySyncState,
-                    onBack = {
-                        onBackClickCount.incrementAndGet()
-                    },
                     onItemClick = {
                         onItemClickCount.incrementAndGet()
                     },