+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/account/view/AccountView.kt
@@ -2,12 +2,9 @@ package co.electriccoin.zcash.ui.screen.account.view
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
@@ -19,42 +16,76 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import cash.z.ecc.android.sdk.Synchronizer
+import cash.z.ecc.android.sdk.fixture.TransactionOverviewFixture
+import cash.z.ecc.android.sdk.model.TransactionOverview
+import cash.z.ecc.android.sdk.model.Zatoshi
 import co.electriccoin.zcash.ui.R
 import co.electriccoin.zcash.ui.common.compose.BalanceWidget
 import co.electriccoin.zcash.ui.common.compose.DisableScreenTimeout
 import co.electriccoin.zcash.ui.common.model.WalletSnapshot
 import co.electriccoin.zcash.ui.common.test.CommonTag
-import co.electriccoin.zcash.ui.design.MINIMAL_WEIGHT
 import co.electriccoin.zcash.ui.design.component.GradientSurface
-import co.electriccoin.zcash.ui.design.component.PrimaryButton
 import co.electriccoin.zcash.ui.design.component.SmallTopAppBar
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
 import co.electriccoin.zcash.ui.fixture.WalletSnapshotFixture
 import co.electriccoin.zcash.ui.screen.account.AccountTag
+import co.electriccoin.zcash.ui.screen.account.state.TransactionHistorySyncState
+import kotlinx.collections.immutable.persistentListOf
 
-@Preview("Account")
+@Preview("Account No History")
 @Composable
-private fun ComposablePreview() {
+private fun HistoryLoadingComposablePreview() {
     ZcashTheme(forceDarkMode = false) {
         GradientSurface {
             Account(
                 walletSnapshot = WalletSnapshotFixture.new(),
                 isKeepScreenOnWhileSyncing = false,
-                goHistory = {},
                 goBalances = {},
                 goSettings = {},
+                transactionState = TransactionHistorySyncState.Loading,
+                onItemClick = {},
+                onTransactionIdClick = {}
             )
         }
     }
 }
 
 @Composable
+@Preview("Account History List")
+private fun HistoryListComposablePreview() {
+    ZcashTheme(forceDarkMode = false) {
+        GradientSurface {
+            Account(
+                walletSnapshot = WalletSnapshotFixture.new(),
+                isKeepScreenOnWhileSyncing = false,
+                goBalances = {},
+                goSettings = {},
+                transactionState =
+                    TransactionHistorySyncState.Syncing(
+                        @Suppress("MagicNumber")
+                        persistentListOf(
+                            TransactionOverviewFixture.new(netValue = Zatoshi(100000000)),
+                            TransactionOverviewFixture.new(netValue = Zatoshi(200000000)),
+                            TransactionOverviewFixture.new(netValue = Zatoshi(300000000)),
+                        )
+                    ),
+                onItemClick = {},
+                onTransactionIdClick = {}
+            )
+        }
+    }
+}
+
+@Composable
+@Suppress("LongParameterList")
 fun Account(
-    walletSnapshot: WalletSnapshot,
-    isKeepScreenOnWhileSyncing: Boolean?,
     goBalances: () -> Unit,
-    goHistory: () -> Unit,
     goSettings: () -> Unit,
+    isKeepScreenOnWhileSyncing: Boolean?,
+    onItemClick: (TransactionOverview) -> Unit,
+    onTransactionIdClick: (String) -> Unit,
+    transactionState: TransactionHistorySyncState,
+    walletSnapshot: WalletSnapshot,
 ) {
     Scaffold(topBar = {
         AccountTopAppBar(onSettings = goSettings)
@@ -62,14 +93,15 @@ fun Account(
         AccountMainContent(
             walletSnapshot = walletSnapshot,
             isKeepScreenOnWhileSyncing = isKeepScreenOnWhileSyncing,
-            goHistory = goHistory,
             goBalances = goBalances,
+            transactionState = transactionState,
+            onItemClick = onItemClick,
+            onTransactionIdClick = onTransactionIdClick,
             modifier =
                 Modifier.padding(
                     top = paddingValues.calculateTopPadding() + ZcashTheme.dimens.spacingDefault,
-                    bottom = paddingValues.calculateBottomPadding() + ZcashTheme.dimens.spacingHuge,
-                    start = ZcashTheme.dimens.screenHorizontalSpacingRegular,
-                    end = ZcashTheme.dimens.screenHorizontalSpacingRegular
+                    // We intentionally do not set the bottom and horizontal paddings here. Those are set by the
+                    // underlying transaction history composable
                 )
         )
     }
@@ -94,34 +126,37 @@ private fun AccountTopAppBar(onSettings: () -> Unit) {
 }
 
 @Composable
+@Suppress("LongParameterList")
 private fun AccountMainContent(
     walletSnapshot: WalletSnapshot,
     isKeepScreenOnWhileSyncing: Boolean?,
     goBalances: () -> Unit,
-    goHistory: () -> Unit,
+    onItemClick: (TransactionOverview) -> Unit,
+    onTransactionIdClick: (String) -> Unit,
+    transactionState: TransactionHistorySyncState,
     modifier: Modifier = Modifier
 ) {
     Column(
-        Modifier
-            .fillMaxHeight()
-            .verticalScroll(rememberScrollState())
-            .then(modifier),
+        modifier = modifier,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
-
-        BalancesStatus(walletSnapshot, goBalances)
-
-        Spacer(
-            modifier =
-                Modifier
-                    .fillMaxHeight()
-                    .weight(MINIMAL_WEIGHT)
-        )
-
         Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingSmall))
 
-        PrimaryButton(onClick = goHistory, text = stringResource(R.string.account_button_history))
+        BalancesStatus(
+            walletSnapshot = walletSnapshot,
+            goBalances = goBalances,
+            modifier =
+                Modifier
+                    .padding(horizontal = ZcashTheme.dimens.screenHorizontalSpacingRegular)
+        )
+
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingXlarge))
+
+        HistoryContainer(
+            transactionState = transactionState,
+            onItemClick = onItemClick,
+            onTransactionIdClick = onTransactionIdClick,
+        )
 
         if (isKeepScreenOnWhileSyncing == true && walletSnapshot.status == Synchronizer.Status.SYNCING) {
             DisableScreenTimeout()
@@ -130,16 +165,18 @@ private fun AccountMainContent(
 }
 
 @Composable
-@Suppress("LongMethod")
 private fun BalancesStatus(
     walletSnapshot: WalletSnapshot,
-    goBalances: () -> Unit
+    goBalances: () -> Unit,
+    modifier: Modifier = Modifier
 ) {
     Column(
         modifier =
-            Modifier
-                .fillMaxWidth()
-                .testTag(AccountTag.BALANCE_VIEWS),
+            modifier.then(
+                Modifier
+                    .fillMaxWidth()
+                    .testTag(AccountTag.BALANCE_VIEWS)
+            ),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
         BalanceWidget(